jobber/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-multipart
pydantic[email]
sqlalchemy
PyJWT